+++ libraries/MCP23017_RT/CHANGELOG.md
@@ -6,11 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.6.3] 2024-05-25
+- sync MCP23S17, add several interrupt functions (experimental)
+- add **MCP23x17_registers.h**
+- add **readReg16()** and **writeReg16()** optimization
+- update readme.md
+- update keywords.txt
+
 ## [0.6.2] 2024-03-02
 - Fix #29, add parameter to **begin(bool pullup)**
 - remove DATE field from examples as it adds no value.
 - update GitHub/actions to version v4 in workflows.
-
 
 ## [0.6.1] 2024-01-05
 - Fix URL in examples